refactor(assumptions): migrate ntheory predicates to new design
Migrate Q.odd
